requirements.txt
pdfplumber
requests
beautifulsoup4
python-dotenv
lxml